refactor(ui): TMS-2508 change roles to view
The linter was complaining that Roles isn't a valid name so renamed it to RolesView and set it up as such.
